# openapi.yaml
openapi: 3.1.0
info:
  title: Pong API
  description: |
    Pong is a health monitoring API that lets you track uptime, latency, and
    availability of your services. Create monitors for HTTP endpoints, SSL
    certificates, heartbeat checks, keyword presence, and port
    availability. Configure alert contacts to get notified when things go wrong,
    and publish public status pages so your users stay informed.

    ## Authentication

    All endpoints under `/api/v1/` require authentication via one of:

    - **Auth0 JWT** — RS256 Bearer token issued by Auth0 (used by the web app)
    - **API Key** — Long-lived key in the format `pong_<hex>` (used for scripts and CI/CD)

    Both are passed as `Authorization: Bearer <token>`.

    API keys can be created and managed via the `/api/v1/api-keys` endpoints.
    The raw key is only shown once at creation time.

    Public endpoints such as heartbeat pings and status pages do not require
    authentication.
  version: 1.0.0
servers:
  - url: https://api.getpong.dev
    description: Production
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 RS256 JSON Web Token
  schemas:
    Monitor:
      type: object
      description: A monitor that periodically checks the health of a target.
      required:
        - id
        - user_id
        - name
        - type
        - target
        - interval_secs
        - timeout_secs
        - confirmation_count
        - consecutive_fails
        - enabled
        - status
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          description: Unique monitor identifier.
          example: 42
        user_id:
          type: integer
          description: ID of the user who owns this monitor.
          example: 1
        name:
          type: string
          description: Human-readable name for the monitor.
          example: "Production API"
        type:
          type: string
          enum: [http, ssl, heartbeat, keyword, port]
          description: The kind of check this monitor performs.
          example: "http"
        target:
          type: string
          description: URL, hostname, or address to monitor.
          example: "https://api.example.com/healthz"
        interval_secs:
          type: integer
          description: Seconds between checks.
          example: 300
        timeout_secs:
          type: integer
          description: Seconds before a check is considered timed out.
          example: 30
        keyword:
          type: string
          nullable: true
          description: Keyword to search for in the response body (keyword monitors only).
          example: "ok"
        keyword_type:
          type: string
          nullable: true
          enum: [contains, not_contains]
          description: Whether the keyword should or should not appear in the response.
          example: "contains"
        keyword_match:
          type: string
          nullable: true
          enum: [exact, regex]
          description: Match mode for keyword checks. "exact" for plain string, "regex" for regular expression.
          default: exact
          example: "exact"
        expected_status:
          type: integer
          nullable: true
          description: Expected HTTP status code (HTTP monitors).
          example: 200
        latency_warn_ms:
          type: integer
          nullable: true
          description: Latency threshold in milliseconds that triggers a warning.
          example: 1000
        confirmation_count:
          type: integer
          description: Number of consecutive failures required before marking the monitor as down.
          example: 1
        consecutive_fails:
          type: integer
          description: Current count of consecutive check failures.
          example: 0
        heartbeat_token:
          type: string
          nullable: true
          description: Unique token for heartbeat monitors. Generated by the backend.
          example: "hb_a1b2c3d4e5f6"
        heartbeat_last_ping:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the last heartbeat ping received.
          example: "2026-03-29T10:15:00Z"
        heartbeat_secret:
          type: string
          nullable: true
          description: Optional secret for heartbeat monitors. If set, pings must include the X-Secret header with this value.
          example: "a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
        ssl_warn_days:
          type: integer
          nullable: true
          description: Number of days before SSL expiry to trigger a warning.
          example: 30
        ssl_expiry_at:
          type: string
          format: date-time
          nullable: true
          description: SSL certificate expiry timestamp.
          example: "2026-12-01T00:00:00Z"
        enabled:
          type: boolean
          description: Whether the monitor is actively running checks.
          example: true
        status:
          type: string
          enum: [up, down, unknown]
          description: Current status of the monitored target.
          example: "up"
        last_checked_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of the most recent check.
          example: "2026-03-29T12:00:00Z"
        created_at:
          type: string
          format: date-time
          description: When the monitor was created.
          example: "2026-01-15T08:30:00Z"
        updated_at:
          type: string
          format: date-time
          description: When the monitor was last updated.
          example: "2026-03-29T12:00:00Z"
        alert_contact_ids:
          type: array
          nullable: true
          items:
            type: integer
          description: IDs of alert contacts to notify when the monitor changes state.
          example: [1, 3]
    MonitorCreate:
      type: object
      description: Request body for creating a new monitor.
      required:
        - name
        - type
      properties:
        name:
          type: string
          description: Human-readable name for the monitor.
          example: "Production API"
        type:
          type: string
          enum: [http, ssl, heartbeat, keyword, port]
          description: The kind of check this monitor performs.
          example: "http"
        target:
          type: string
          description: URL, hostname, or address to monitor. Not required for heartbeat monitors.
          example: "https://api.example.com/healthz"
        interval_secs:
          type: integer
          description: Seconds between checks.
          default: 300
          example: 300
        timeout_secs:
          type: integer
          description: Seconds before a check is considered timed out.
          default: 30
          example: 30
        keyword:
          type: string
          description: Keyword to search for (keyword monitors only).
          example: "ok"
        keyword_type:
          type: string
          enum: [contains, not_contains]
          description: Whether the keyword should or should not appear.
          example: "contains"
        keyword_match:
          type: string
          enum: [exact, regex]
          description: Match mode for keyword checks. "exact" for plain string, "regex" for regular expression.
          default: exact
          example: "exact"
        expected_status:
          type: integer
          description: Expected HTTP status code.
          default: 200
          example: 200
        latency_warn_ms:
          type: integer
          description: Latency warning threshold in milliseconds.
          example: 1000
        confirmation_count:
          type: integer
          description: Consecutive failures required before marking as down.
          default: 1
          example: 1
        ssl_warn_days:
          type: integer
          description: Days before SSL expiry to warn.
          default: 30
          example: 30
        heartbeat_secret:
          type: string
          description: Optional secret for heartbeat monitors. If set, pings must include the X-Secret header with this value.
          example: "a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
        alert_contact_ids:
          type: array
          items:
            type: integer
          description: IDs of alert contacts to attach.
          example: [1, 3]
    MonitorUpdate:
      type: object
      description: Request body for updating an existing monitor. All fields are optional.
      properties:
        name:
          type: string
          description: Human-readable name for the monitor.
          example: "Staging API"
        type:
          type: string
          enum: [http, ssl, heartbeat, keyword, port]
          description: The kind of check this monitor performs.
          example: "http"
        target:
          type: string
          description: URL, hostname, or address to monitor.
          example: "https://staging.example.com/healthz"
        interval_secs:
          type: integer
          description: Seconds between checks.
          example: 60
        timeout_secs:
          type: integer
          description: Seconds before a check is considered timed out.
          example: 10
        keyword:
          type: string
          description: Keyword to search for (keyword monitors only).
          example: "healthy"
        keyword_type:
          type: string
          enum: [contains, not_contains]
          description: Whether the keyword should or should not appear.
          example: "contains"
        keyword_match:
          type: string
          enum: [exact, regex]
          description: Match mode for keyword checks. "exact" for plain string, "regex" for regular expression.
          example: "exact"
        expected_status:
          type: integer
          description: Expected HTTP status code.
          example: 200
        latency_warn_ms:
          type: integer
          description: Latency warning threshold in milliseconds.
          example: 500
        confirmation_count:
          type: integer
          description: Consecutive failures required before marking as down.
          example: 3
        ssl_warn_days:
          type: integer
          description: Days before SSL expiry to warn.
          example: 14
        heartbeat_secret:
          type: string
          description: Optional secret for heartbeat monitors. If set, pings must include the X-Secret header with this value.
          example: "a3f8b2c1d4e5f6a7b8c9d0e1f2a3b4c5"
        alert_contact_ids:
          type: array
          items:
            type: integer
          description: IDs of alert contacts to attach.
          example: [2]
    CheckResult:
      type: object
      description: The result of a single monitor check.
      required:
        - id
        - monitor_id
        - status
        - latency_ms
        - region
        - checked_at
      properties:
        id:
          type: integer
          description: Unique check result identifier.
          example: 98765
        monitor_id:
          type: integer
          description: ID of the monitor this result belongs to.
          example: 42
        status:
          type: string
          enum: [up, down]
          description: Result of the check.
          example: "up"
        latency_ms:
          type: integer
          description: Response latency in milliseconds.
          example: 132
        status_code:
          type: integer
          nullable: true
          description: HTTP status code returned (if applicable).
          example: 200
        message:
          type: string
          nullable: true
          description: Additional details or error message.
          example: "OK"
        region:
          type: string
          description: Geographic region from which the check was performed.
          example: "eu-west-1"
        checked_at:
          type: string
          format: date-time
          description: When the check was executed.
          example: "2026-03-29T12:00:00Z"
...
    DailyUptime:
      type: object
      description: Uptime data for a single day.
      required:
        - date
        - percentage
        - total_checks
      properties:
        date:
          type: string
          format: date
          description: The date (YYYY-MM-DD).
          example: "2026-03-28"
        percentage:
          type: number
          format: double
          description: Uptime percentage for the day (0-100).
          example: 99.95
        total_checks:
          type: integer
          description: Total number of checks performed that day.
          example: 288
...
    Error:
      type: object
      description: Error response returned when a request fails.
      required:
        - error
      properties:
        error:
          type: string
          description: A human-readable error message.
          example: "monitor not found"
  parameters:
    MonitorId:
      name: id
      in: path
      required: true
      description: Unique monitor identifier.
      schema:
        type: integer
        example: 42
...
  responses:
    Unauthorized:
      description: Authentication credentials are missing or invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "unauthorized"
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "not found"
    BadRequest:
      description: The request body is invalid or missing required fields.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "invalid request body"
paths:
...
  /api/v1/monitors:
    get:
      operationId: listMonitors
      tags: [Monitors]
      summary: List monitors
      description: Returns all monitors belonging to the authenticated user.
      responses:
        "200":
          description: A list of monitors.
          content:
            application/json:
              schema:
                type: object
                required: [monitors]
                properties:
                  monitors:
                    type: array
                    items:
                      $ref: "#/components/schemas/Monitor"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      operationId: createMonitor
      tags: [Monitors]
      summary: Create a monitor
      description: |
        Creates a new monitor for the authenticated user.

        For **heartbeat** monitors the `target` field is not required; the
        backend will generate a unique `heartbeat_token` that can be used to
        ping the heartbeat endpoint.

        Default values when not specified:
        - `interval_secs`: 300
        - `timeout_secs`: 30
        - `expected_status`: 200
        - `confirmation_count`: 1
        - `ssl_warn_days`: 30
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorCreate"
            examples:
              http_monitor:
                summary: HTTP monitor
                value:
                  name: "Production API"
                  type: "http"
                  target: "https://api.example.com/healthz"
                  interval_secs: 60
                  expected_status: 200
                  alert_contact_ids: [1]
              heartbeat_monitor:
                summary: Heartbeat monitor
                value:
                  name: "Cron Job Heartbeat"
                  type: "heartbeat"
                  interval_secs: 3600
              keyword_monitor:
                summary: Keyword monitor
                value:
                  name: "Homepage keyword check"
                  type: "keyword"
                  target: "https://example.com"
                  keyword: "Welcome"
                  keyword_type: "contains"
      responses:
        "201":
          description: Monitor created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/monitors/{id}:
    get:
      operationId: getMonitor
      tags: [Monitors]
      summary: Get a monitor
      description: Returns a single monitor by ID.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "200":
          description: The requested monitor.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      operationId: updateMonitor
      tags: [Monitors]
      summary: Update a monitor
      description: Updates an existing monitor. Only the fields provided in the request body are changed.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MonitorUpdate"
      responses:
        "200":
          description: Monitor updated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Monitor"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    delete:
      operationId: deleteMonitor
      tags: [Monitors]
      summary: Delete a monitor
      description: Permanently deletes a monitor and all its associated check results.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
      responses:
        "204":
          description: Monitor deleted successfully.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/monitors/{id}/results:
    get:
      operationId: listCheckResults
      tags: [Monitors]
      summary: List check results
      description: Returns paginated check results for a monitor, ordered by most recent first.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: limit
          in: query
          description: Maximum number of results to return.
          schema:
            type: integer
            default: 50
            minimum: 1
            maximum: 1000
            example: 50
        - name: offset
          in: query
          description: Number of results to skip for pagination.
          schema:
            type: integer
            default: 0
            minimum: 0
            example: 0
      responses:
        "200":
          description: A paginated list of check results.
          content:
            application/json:
              schema:
                type: object
                required: [results, total, limit, offset]
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/CheckResult"
                  total:
                    type: integer
                    description: Total number of results available.
                    example: 1440
                  limit:
                    type: integer
                    description: Maximum results returned per page.
                    example: 50
                  offset:
                    type: integer
                    description: Number of results skipped.
                    example: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/monitors/{id}/uptime/daily:
    get:
      operationId: getDailyUptime
      tags: [Monitors]
      summary: Get daily uptime
      description: Returns day-by-day uptime data for a monitor over a specified number of days.
      parameters:
        - $ref: "#/components/parameters/MonitorId"
        - name: days
          in: query
          description: Number of days of daily uptime data to return.
          schema:
            type: integer
            default: 90
            minimum: 1
            example: 90
      responses:
        "200":
          description: Daily uptime breakdown.
          content:
            application/json:
              schema:
                type: object
                required: [monitor_id, days, daily]
                properties:
                  monitor_id:
                    type: integer
                    description: ID of the monitor.
                    example: 42
                  days:
                    type: integer
                    description: Number of days covered.
                    example: 90
                  daily:
                    type: array
                    items:
                      $ref: "#/components/schemas/DailyUptime"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ---------- Alert Contacts ----------
...